Add files via upload
Travel agencies have a knowledge base of the available offers they could recommend to customers. For
each of those offers, the accommodation and means of transportation is specified. Moreover, they all also
have knowledge about a customer’s preferred accommodation, transportation and the type of activities
he/she would like to perform. As a result, the system should be provided with this data to be able to
recommend to the customer an offer accordingly.
For the system to recommend an offer to the customer, he/she should provide a list of their preferences
indicating, for instance the activities he/she would like to perform, the type of accommodation he/she
would like to stay in, etc. The system tries then to satisfy as much of those preferences as much as possible
and recommend an offer accordingly.
The system could also recommend an offer to a group of people, as each offer could be scheduled to a
maximum of N guests. Each customer will provide his list of preferences and the system will try to choose
a set of preferences that is satisfied by an offer. It will then pick maximum N customers who will be most
satisfied by the offer(based on the data present in the knowledge base).